Add `build_helper` crate to share code between tidy and bootstrap
